messenger protocol changes
There is now a plaintext banner that is exchanged so that both ends can
verify they are talking to the right software.  This string should be
changed each time the wire protocol changes.

There are also mon, osd, and mds specific protocol versions defined, but
these are not yet used.