osd: send and process heartbeats in separate thread, channel
Use a separate dispatch thread to process heartbeats.  Use a
separate thread to send them.  This ensures something slow
(e.g. a map update) does not make an osd appear to be down.

This also means a spearate entity_addr for heartbeats, which puts
them over a separate TCP stream.