Makefile: link libcommon, lockdep last
Link libcommon.a last for everything.

Include lockdep last in libcommon so to initialize first, shutdown
last, so that g_lockdep is initialized to zero for all Mutexes.

Put config.cc next to last so that the atomic_t thread counter is
initialized before all threads are created, shutdown after all
threads are destroyed.