kclient: locally open file based on what MDS thinks we want, not what we still want
As long as we have our interest in the caps still registered with
the MDS, we can do a local open.  Checking what we currently still
want is only helpful if the same file still is open by another
process or something.  Now, we can do a local open on a close +
wait + reopen, as long as we haven't done a cap release yet.