kclient: do readdir from dcache when possible
If we have I_COMPLETE, do a readdir straight from the dcache.  When
we do a readdir prepopulate, reorder parent->d_subdirs to match the
mds order, so that the dcache readdir order matches.

Some wonkiness here to detect the end of directory.  There is also
some inconsistency if the directory changes mid-readdir and we
have to switch back to talking to the mds.