kclient: avoid d_time reuse s.t. leases and dir rdcache can coexist
We want to be able to remember a lease even when we are issued
RDCACHE on the entire dir.  (Otherwise, we fail to release the
dentry when doing an op, and have to wait for the revoke round
trip.)

So avoid reusing d_time for the rdcache_gen.  Unconditionally set
it in update_dentry_lease.

Also, force inclusion of a no-op inode update if we need to include
the dentry lease release, even when the dir caps aren't changing.