kclient: fix readdir vs rm
Okay, do not rely on MDS to provide dentry positioning information,
since it is all relative to the start _string_ we provide, and that
can change directory position without notice.

Simplify readdir a bit wrt seeks.  A seek to 0, a new frag, or
prior to the current chunk resets buffered state.

For each frag, we walk through chunks, always in order.  We set
dentry positions/offsets based on the frag and position within our
sweep across the frag.  Successive chunks are grabbed from the MDS
relative to a filename (not offset), so concurrent
insertions/removals don't bother us (although we will not see
insertions lexicographically prior to our position).