osd: use per-PG ObjectStore::Sequencer when possible
Otherwise, when it doesn't matter, use the default.  This
serialized io submission to btrfs at the lower layer within a
single PG.  In general this is not significant, since we have
lots of PGs.  We could potentionally do per object in some cases,
but we'd have to be careful about snapshots and such.  Keep it
simple for now.