Revert "cfuse.cc: use safe_write"
This reverts commit e8ac5aa2a4c4e3ce84ed553dbebfb1cccf5679a9.

This commit is just erroneous. It adds checks on a pipe write
for the result and an abort if the write failed. But that's broken
in the desired case where we succeed, block on ceph_fuse_ll_main(),
and the parent process is long-gone by the time we get to this code!

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>