cls_rgw: implement a dir_suggest_changes function.
This takes a bufferlist of suggested changes to the directory, trims
out any sufficiently old tags, and then applies the suggested changes
to any entry which has no remaining tags.
Right now "sufficiently old" is controlled by a #define. Obviously
this should be made into a config member if possible; I need to check
out the possibilities for that when you're in a class.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>