ceph-kdump-copy: add tools for saving kdumps
This puts in place an init script and a command it runs to save a
kernel core dump to a remote server when a panic or other kernel
failure occurs.  The content of a crash file generated by the
"kdump" init script (which uses the Ubuntu "apport" kernel_crashdump
command) is re-packaged and copied (via scp) to a remote server
specified by variables in /etc/default/ceph-kdump-copy.  This
packaging (as well as the work done by the apport script) is done
on the next boot of the kernel following a crash.


Note:  Although the init script, its config file, and shell script
are known to work, the packaging of these things should not be
expected to be.  I (now) know, for example, that the init script
belongs in src/init-ceph-kdump-copy.in if it is to follow the
pattern used by the ceph init script.  And there are likely other
problems with the build/install portions of this patch.

Signed-off-by: Alex Elder <elder@dreamhost.com>