ceph: add partial to validate()
Allow validation of the front of a command, to precalculate the
target for a command (basically, whether it's a 'normal' monitor
command, a "tell <osd>" command, or a "pg <pgid>" command; the latter
two go to the OSD, and so the OSD needs to be consulted for valid
command lists *and* receive the command directly.  This means we
have to have some fixed preclassification of the command, which
is done with validate(partial=True).

Signed-off-by: Dan Mick <dan.mick@inktank.com>